Various fixes (#668)
* Fixed warnings about signed/unsigned mismatch by changing to unsigned a few values that are only ever assigned or compared to unsigned values, yet were signed for some reason.
* Fixed TextEntryPanel's event handler missing the check about whether it was the active panel. Among other annoyances, it made impossible to do a search & replace operation on a different tab than the first.
* Added some commented-out stuff about the Strife demo conversation script format. It's more for documentation than anything else, which is why it's commented out.

+++ src/Archive/Archive.cpp
@@ -137,7 +137,7 @@ int ArchiveTreeNode::entryIndex(ArchiveEntry* entry, size_t startfrom)
 		{
 			if (entries[a].get() == entry)
 			{
-				entry->index_guess = (int)a;
+				entry->index_guess = a;
 				return (int)a;
 			}
 		}
@@ -146,7 +146,7 @@ int ArchiveTreeNode::entryIndex(ArchiveEntry* entry, size_t startfrom)
 		{
 			if (entries[a].get() == entry)
 			{
-				entry->index_guess = (int)a;
+				entry->index_guess = a;
 				return (int)a;
 			}
 		}
@@ -154,7 +154,7 @@ int ArchiveTreeNode::entryIndex(ArchiveEntry* entry, size_t startfrom)
 		{
 			if (entries[a].get() == entry)
 			{
-				entry->index_guess = (int)a;
+				entry->index_guess = a;
 				return (int)a;
 			}
 		}
